Move Config.xxx_value helpers to fleunt/config/types.rb
