+++ lib/fluent/config/types.rb
@@ -1,4 +1,49 @@
 module Fluent
+  require 'json'
+
+  module Config
+    def self.size_value(str)
+      case str.to_s
+      when /([0-9]+)k/i
+        $~[1].to_i * 1024
+      when /([0-9]+)m/i
+        $~[1].to_i * (1024 ** 2)
+      when /([0-9]+)g/i
+        $~[1].to_i * (1024 ** 3)
+      when /([0-9]+)t/i
+        $~[1].to_i * (1024 ** 4)
+      else
+        str.to_i
+      end
+    end
+
+    def self.time_value(str)
+      case str.to_s
+      when /([0-9]+)s/
+        $~[1].to_i
+      when /([0-9]+)m/
+        $~[1].to_i * 60
+      when /([0-9]+)h/
+        $~[1].to_i * 60 * 60
+      when /([0-9]+)d/
+        $~[1].to_i * 24 * 60 * 60
+      else
+        str.to_f
+      end
+    end
+
+    def self.bool_value(str)
+      case str.to_s
+      when 'true', 'yes'
+        true
+      when 'false', 'no'
+        false
+      else
+        nil
+      end
+    end
+  end
+
   Configurable.register_type(:string, Proc.new { |val, opts|
     val
   })
